requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
notion-client==2.2.1
python-dotenv==1.0.1
apify-client==1.7.1
lxml==5.1.0
